requirements.txt.txt
pyttsx3
SpeechRecognition
wikipedia-api
webbrowser
smtplib
